Rename InputField to TypedFormField
