+++ src/components/field-components/index.js
@@ -1,7 +1,7 @@
 import CheckboxField from './CheckboxField.vue'
 import DateTimeField from './DateTimeField.vue'
 import FileField from './FileField.vue'
-import InputField from './InputField.vue'
+import TypedFormField from './TypedFormField.vue'
 import RadiosField from './RadiosField.vue'
 import SelectField from './SelectField.vue'
 import TextAreaField from './TextAreaField.vue'
@@ -10,8 +10,8 @@ export default {
   CheckboxField,
   DateTimeField,
   FileField,
-  InputField,
   RadiosField,
   SelectField,
-  TextAreaField
+  TextAreaField,
+  TypedFormField
 }
